requirements.txt
jax==0.4.26
tensorflow
sentencepiece
ml_collections
Pillow